swee it works, the deed is done
Auto uses both patrol and seek.
How to use patrol,
The patrol points needs to be -1.03 y axis as thats our plane right now
Detect radius is how far the enemy can detect the player.
Increasing the enemy agent radius to 0.5 make the patrol smoother.
Lose radius is when the enemy lose sight of the player